added policy directory
